python_modules/libraries/dagster-gcp/requirements.txt
google-cloud-bigquery[pandas,pyarrow]==1.11.2
google-cloud-storage